Group-1/requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
scikit-learn

gradio
sqlite3
bcrypt
torch
transformers
python-dotenv
openai
fpdf
wandb
optuna
huggingface-hub

bcrypt
sqlite3
openai~=1.60.2
python-dotenv~=1.0.1
datasets~=3.2.0
pandas~=2.2.3
tiktoken~=0.8.0
numpy~=2.2.2